gprs_ns2_vc_fsm: fix nullpointer dereference when sending uintdata
the function gprs_ns2_vc_rx() is called from gprs_ns2.c with tp=NULL.
This leads into a null pointer dereference because gprs_ns2_vc_rx() is
accessing tp several times. tp is also handed up into the FSM, with the
exception of unitdata. Apparently for unitdata tp does not exist, so we
can move up the part where unitdata is handed up into the FSM and after
that we can check if tp is NULL to make sure the code which is accessing
it is not executed.

Change-Id: I7d7c95604ba4af4ed4b6019f1d432970225f8d7a