d/control: libosmocore-dev: depend on libmnl-dev
Let libosmocore-dev depend on libmnl-dev, as we build the debian package
without --disable-libmnl and so libosmocore.pc lists libmnl in requires.
This patch should fix the following error when trying to build any
debien package depending on libosmocore-dev:

  Package 'libmnl', required by 'libosmocore', not found

A similar change in the rpm spec is not be needed, as rpm adds depends
from pc files automatically.

Fixes: 4eb89afa ("configure --enable-libmnl: Add libmnl to libosmocore.pc.in Requires")
Change-Id: I7fd578fcc22c0ea6ac04bdc26dc1741e4e26767c